Issue #4365: Split and Organize Checkstyle inputs by Test for checks in header package
