--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/InputRegexpHeader3.java
@@ -1,7 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.header;
-
-import java.awt.*;
-
-public class InputRegexpHeader3
-{
-}
