--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/InputRegexpHeader6.java
@@ -1,3 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.header;
-
-import java.awt.*;
