--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/InputRegexpSmallHeader.java
@@ -1,6 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.header;
-
-
-/**
- */
-public class InputRegexpSmallHeader {}
