+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/regexpheader/InputRegexpHeader1.java
@@ -1,0 +1,11 @@
+package com.puppycrawl.tools.checkstyle.checks.header.regexpheader;
+
+import java.awt.*;
+
+/**
+ * Some doc.
+ */
+
+public class InputRegexpHeader1
+{
+}
