+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/regexpheader/InputRegexpHeader3.java
@@ -1,0 +1,7 @@
+package com.puppycrawl.tools.checkstyle.checks.header.regexpheader;
+
+import java.awt.*;
+
+public class InputRegexpHeader3
+{
+}
