+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/regexpheader/InputRegexpHeader4.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.header;
+package com.puppycrawl.tools.checkstyle.checks.header.regexpheader;
 
 import java.awt.*;
 import java.awt.*;
