+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/regexpheader/InputRegexpHeader5.java
@@ -1,5 +1,5 @@
 //////////////
-package com.puppycrawl.tools.checkstyle.checks.header;
+package com.puppycrawl.tools.checkstyle.checks.header.regexpheader;
 
 import java.awt.*;
 import java.awt.*;
