+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/regexpheader/InputRegexpHeader6.java
@@ -1,0 +1,3 @@
+package com.puppycrawl.tools.checkstyle.checks.header.regexpheader;
+
+import java.awt.*;
