+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/regexpheader/InputRegexpSmallHeader.java
@@ -1,0 +1,6 @@
+package com.puppycrawl.tools.checkstyle.checks.header.regexpheader;
+
+
+/**
+ */
+public class InputRegexpSmallHeader {}
