+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/regexpheader/regexp.header
