+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/header/regexpheader/regexp2.header4
